@@ -1,6 +1,6 @@
 {
   "name": "node-poweredup",
-  "version": "3.5.1",
+  "version": "4.2.0",
   "description": "A Javascript module to interface with LEGO Powered Up components.",
   "homepage": "https://github.com/nathankellenicki/node-poweredup/",
   "main": "dist/node/index-node.js",
@@ -16,26 +16,21 @@
   "author": "Nathan Kellenicki <nathan@kellenicki.com>",
   "license": "MIT",
   "dependencies": {
-    "compare-versions": "^3.5.0",
+    "compare-versions": "^3.5.1",
     "debug": "^4.1.1",
-    "noble": "1.9.1",
-    "noble-mac": "git+https://github.com/Timeular/noble-mac.git#af4418e"
+    "@abandonware/noble": "1.9.2-5"
   },
   "devDependencies": {
-    "@types/debug": "4.1.4",
-    "@types/noble": "0.0.39",
-    "@types/node": "^11.13.7",
+    "@types/debug": "4.1.5",
+    "@types/node": "^12.12.7",
     "@types/web-bluetooth": "0.0.4",
     "ink-docstrap": "^1.3.2",
     "jsdoc": "^3.6.3",
-    "jsdoc-to-markdown": "^5.0.0",
-    "ts-loader": "^5.4.3",
-    "tslint": "^5.16.0",
-    "typescript": "^3.4.5",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.1"
-  },
-  "resolutions": {
-    "xpc-connection": "sandeepmistry/node-xpc-connection#pull/26/head"
+    "jsdoc-to-markdown": "^5.0.2",
+    "ts-loader": "^6.2.1",
+    "tslint": "^5.20.1",
+    "typescript": "^3.7.2",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10"
   }
 }
